Refactored several data models into one, TraceModel. Defined Trace Data abstraction. TraceHelper is operational
