+++ src/traceService/es-analyzer.js
@@ -1,89 +1,22 @@
 import esprima from 'esprima';
 import estraverse from 'estraverse';
+
 export class EsAnalyzer {
 
-  constructor() {
-    this.esprima = esprima;
-    this.estraverse = estraverse;
-    this.init();
-  }
-  
-  init(){
-        let Syntax = {
-        AssignmentExpression: 'AssignmentExpression',
-        ArrayExpression: 'ArrayExpression',
-        BlockStatement: 'BlockStatement',
-        BinaryExpression: 'BinaryExpression',
-        BreakStatement: 'BreakStatement',
-        CallExpression: 'CallExpression',
-        CatchClause: 'CatchClause',
-        ConditionalExpression: 'ConditionalExpression',
-        ContinueStatement: 'ContinueStatement',
-        DoWhileStatement: 'DoWhileStatement',
-        DebuggerStatement: 'DebuggerStatement',
-        EmptyStatement: 'EmptyStatement',
-        ExpressionStatement: 'ExpressionStatement',
-        ForStatement: 'ForStatement',
-        ForInStatement: 'ForInStatement',
-        FunctionDeclaration: 'FunctionDeclaration',
-        FunctionExpression: 'FunctionExpression',
-        Identifier: 'Identifier',
-        IfStatement: 'IfStatement',
-        Literal: 'Literal',
-        LabeledStatement: 'LabeledStatement',
-        LogicalExpression: 'LogicalExpression',
-        MemberExpression: 'MemberExpression',
-        NewExpression: 'NewExpression',
-        ObjectExpression: 'ObjectExpression',
-        Program: 'Program',
-        Property: 'Property',
-        ReturnStatement: 'ReturnStatement',
-        SequenceExpression: 'SequenceExpression',
-        SwitchStatement: 'SwitchStatement',
-        SwitchCase: 'SwitchCase',
-        ThisExpression: 'ThisExpression',
-        ThrowStatement: 'ThrowStatement',
-        TryStatement: 'TryStatement',
-        UnaryExpression: 'UnaryExpression',
-        UpdateExpression: 'UpdateExpression',
-        VariableDeclaration: 'VariableDeclaration',
-        VariableDeclarator: 'VariableDeclarator',
-        WhileStatement: 'WhileStatement',
-        WithStatement: 'WithStatement'
-    };
-    this.traceTypes = {
-        LocalStack : [Syntax.FunctionDeclaration, Syntax.FunctionExpression],
-        Expression: [
-            Syntax.UnaryExpression,
-            Syntax.UpdateExpression,
-            Syntax.CallExpression,
-            Syntax.Property,
-            Syntax.VariableDeclarator,
-            Syntax.AssignmentExpression,
-            Syntax.BinaryExpression,
-            Syntax.ReturnStatement,
-            Syntax.ForStatement,
-            Syntax.ForInStatement,
-            Syntax.WhileStatement,
-            Syntax.DoWhileStatement,
-            Syntax.ExpressionStatement
-            ],
-        ExpressionStatement : [
-            Syntax.ExpressionStatement
-            ],
-        ControlFlow : [],
-        Condition: [],
-        Loop: [Syntax.WhileStatement],
-        exception: []
+    constructor(traceModel) {
+        this.traceModel = traceModel;
+        this.Syntax = this.traceModel.esSyntax; 
+        this.traceTypes = this.traceModel.traceTypes;
         
-    };
-    this.Syntax = Syntax; 
-  }
-  
+        this.esprima = esprima;
+        this.estraverse = estraverse;
+    
+    }
+
     traverse(object, visitor, master, objectKey) {
         var key, child, parent, path;
 
-        parent = (typeof master === 'undefined') ? [] : master;
+        parent = !master ? [] : master;
 
         if (visitor.call(null, object, parent, objectKey) === false) {
             return;
@@ -136,19 +69,14 @@ export class EsAnalyzer {
         return beautifulString;
     }
     
-
     traceAllAutoLog(code, autoLogTracer) {
- 
-            
             let tree = esprima.parse(code, { range: true, loc: true });
-
-
+            
             this.traverse(tree, function traceVisitor(node, path, nodeKey) {
-                autoLogTracer({'node' :node, 'code' : code, 'path' :path, 'nodeKey' :nodeKey});
+                autoLogTracer({node : node, code : code, path :path, nodeKey :nodeKey});
             });
             
           return tree;
-        
     }
    
 }
