+++ src/traceService/external/esprima-node-factory.js
