+++ src/traceService/trace-helper.js
@@ -1,28 +1,17 @@
+import {TraceModel} from './trace-model';
 
 export class TraceHelper {
-    constructor(executionTrace){
-        this.executionTrace = executionTrace;
-        this.Syntax = {
-            AssignmentExpression: 'AssignmentExpression',
-            BinaryExpression: 'BinaryExpression', 
-            CallExpression: 'CallExpression',
-            DoWhileStatement: 'DoWhileStatement',
-            ExpressionStatement: 'ExpressionStatement', 
-            ForStatement: 'ForStatement', 
-            ForInStatement: 'ForInStatement', 
-            FunctionDeclaration: 'FunctionDeclaration',
-            FunctionExpression: 'FunctionExpression',
-            IfStatement: 'IfStatement', 
-            NewExpression: 'NewExpression',
-            Property: 'Property',
-            ReturnStatement: 'ReturnStatement',
-            UnaryExpression: 'UnaryExpression',
-            UpdateExpression: 'UpdateExpression', 
-            VariableDeclaration: 'VariableDeclaration', 
-            VariableDeclarator: 'VariableDeclarator',
-            WhileStatement: 'WhileStatement'
-        };      
+    constructor(trace){
+        this.traceModel = new TraceModel();
+        this.Syntax = this.traceModel.traceSyntax;
+        this.makeTrace(trace);
     }
+    
+    makeTrace(trace){
+        this.trace = this.traceModel.makeTrace(trace);
+    }
+    
+
     /*
      * getLineValues(lineNumber,valueTable)
      * @ param lineNumber - the line number in the editor for retrieving all the associated values
@@ -243,6 +232,100 @@ export class TraceHelper {
     	
     	return stackText;  
     
+    }
+    
+    autoLog(info) {
+                var key = info.text + ':' + info.indexRange[0]+':' + info.indexRange[1];
+                
+                if(traceTypes.LocalStack.indexOf(info.type)>-1){
+    				this.stack.push(key) ;
+                }
+
+                if(info.type === Syntax.VariableDeclarator || info.type === Syntax.AssignmentExpression){
+                   this.values.push({'id': info.id , 'value': JSON.stringify(info.value), 'range': info.range}); 
+                }
+
+                this.timeline.push({ id: info.id , value: JSON.stringify(info.value), range: info.range, type: info.type, text: info.text});
+
+
+                var stackTop =	this.stack.length - 1;
+                
+				if (this.hits.hasOwnProperty(key)) {
+                    this.hits[key] = this.hits[key] + 1;
+                    this.data[key].hits = this.hits[key] + 1;
+                    this.data[key].values.push({'stackIndex': stackTop, 'value' :JSON.stringify(info.value)});
+                } else {
+                    
+                    if(info.type === Syntax.VariableDeclarator){
+                       this.variables.push({'id': info.id , 'range': info.range});
+                    }
+                    
+                    this.identifiers.push({'id': info.id , 'range': info.range});
+                    
+                    
+                    this.hits[key] = 1;
+                    this.execution.push(key);
+                    this.data[key] = {
+                        'type' : info.type,
+                        'id' : info.id,
+                        'text' : info.text,
+                        'values': [{'stackIndex': stackTop, 'value' :JSON.stringify(info.value)}],
+                        'range': info.range,
+                        'hits' : 1,
+                        'extra' : info.extra
+                    };
+                }
+                
+                if(window.ISCANCELLED){
+                    throw "Trace Cancelled.";
+                }
+                
+                return info.value;
+    }
+    
+    getStackTrace() {
+                var entry,
+                    stackData = [];
+                for (var i in this.stack) {
+                    if (this.stack.hasOwnProperty(i)) {
+                        entry = this.stack[i];
+                        stackData.push({ index: i, text: entry.split(':')[0], range: this.data[entry].range,  count: this.hits[entry]});
+                    }
+                }
+                return stackData;
+    }
+            
+    getExecutionTraceAll() {
+        let result = [];
+        for (let i in this.execution) {
+            let entry = this.execution[i];
+            if (this.data.hasOwnProperty(entry)) {
+                result.push(this.data[entry]);
+            }
+        }
+        return result;
+    }
+    
+    getExpressions() {
+         return {variables : this.identifiers, timeline: this.timeline};
+    }
+    
+    getVariables(){
+        return {variables : this.variables, values: this.values};
+    }
+    
+    getExecutionTrace() {
+        var i, entry, data, stackData = [];
+        for (i in this.execution) {
+            entry = this.execution[i];
+            if (this.data.hasOwnProperty(entry)) {
+                data =this.data[entry];
+                if(traceTypes.Expression.indexOf(data.type) > -1  ){
+                    stackData.push(this.data[entry]);
+                }
+             }
+        }
+        return stackData;
     }
     
 }
